Move config utils to trackster
