+++ client/src/viz/trackster/tracks.js
@@ -6,9 +6,9 @@ import $ from "jquery";
 import { Dataset } from "mvc/dataset/data";
 import { getAppRoot } from "onload/loadConfig";
 import _ from "underscore";
-import config_mod from "utils/config";
 import _l from "utils/localization";
 import bbi from "viz/bbi-data-manager";
+import config_mod from "viz/config";
 import tools_mod from "viz/tools";
 import filters_mod from "viz/trackster/filters";
 import painters from "viz/trackster/painters";
